fix: padding issues when no comments and missing bottom rounded comments on posts
