+++ playfit-mobile/playfit/lib/components/social/post_feed.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:playfit/components/social/post_detail_page.dart';
-import 'package:playfit/components/social/post_card.dart';
+import 'package:playfit/components/social/post_card/post_card.dart';
 
 class PostFeed extends StatefulWidget {
   const PostFeed({super.key});
